requirements.txt
alembic==0.7.4
argh==0.26.1
cairocffi==0.6
CairoSVG==1.0.13
cffi==0.9.2
cssselect==0.9.1
docutils==0.12
Flask==0.10.1
Flask-Login==0.2.9
Flask-Migrate==1.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==0.16
Flask-WeasyPrint==0.5
Flask-WTF==0.9.4
html5lib==0.999
ipython==2.3.1
itsdangerous==0.24
Jinja2==2.7.2
lxml==3.4.3
Mako==1.0.1
MarkupSafe==0.19
pathtools==0.1.2
Pillow==2.8.1
pycparser==2.10
PyPDF2==1.24
Pyphen==0.9.1
python-ldap==2.4.19
PyYAML==3.11
reportlab==3.1.44
six==1.9.0
SQLAlchemy==0.9.4
tcpwatch==1.3.1
tinycss==0.3
uWSGI==2.0.9
watchdog==0.8.3
WeasyPrint==0.23
webunit==1.3.10
Werkzeug==0.10.1
WTForms==1.0.5
